+++ packages/js/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.1] - 2026-09-03
+
 ### Fixed
 
 - `removeParam()` now reaches nested raw params: `removeParam('custom_sort[lang]')`